move `test-commons` module into `tests` module
