+++ parent/pom.xml
@@ -133,6 +133,12 @@
                 <groupId>org.apache.hbase</groupId>
                 <artifactId>hbase-client</artifactId>
                 <version>${hbase.version}</version>
+                <exclusions>
+                    <exclusion>
+                        <groupId>org.slf4j</groupId>
+                        <artifactId>slf4j-log4j12</artifactId>
+                    </exclusion>
+                </exclusions>
             </dependency>
 
             <dependency>
@@ -390,7 +396,7 @@
                     <failOnError>false</failOnError>
                     <quiet>true</quiet>
                     <excludePackageNames>
-                        fr.inria.atlanmod.neoemf.test.commons.models.*
+                        fr.inria.atlanmod.neoemf.tests.models.*
                     </excludePackageNames>
                 </configuration>
             </plugin>
@@ -500,6 +506,9 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-project-info-reports-plugin</artifactId>
                 <version>${project-info-reports-plugin.version}</version>
+                <configuration>
+                    <dependencyLocationsEnabled>false</dependencyLocationsEnabled>
+                </configuration>
             </plugin>
 
             <plugin>
@@ -534,7 +543,7 @@
                     <failOnError>false</failOnError>
                     <quiet>true</quiet>
                     <excludePackageNames>
-                        fr.inria.atlanmod.neoemf.test.commons.models.*
+                        fr.inria.atlanmod.neoemf.tests.models.*
                     </excludePackageNames>
                 </configuration>
                 <reportSets>
