+++ tests/src/main/java/fr/inria/atlanmod/neoemf/tests/AbstractResourceBuilder.java
@@ -9,7 +9,7 @@
  *     Atlanmod INRIA LINA Mines Nantes - initial API and implementation
  */
 
-package fr.inria.atlanmod.neoemf.test.commons;
+package fr.inria.atlanmod.neoemf.tests;
 
 import fr.inria.atlanmod.neoemf.resources.PersistentResource;
