+++ tests/src/main/resources/MapSample.genmodel
@@ -1,0 +1,59 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<genmodel:GenModel xmi:version="2.0" xmlns:xmi="http://www.omg.org/XMI" xmlns:ecore="http://www.eclipse.org/emf/2002/Ecore"
+    xmlns:genmodel="http://www.eclipse.org/emf/2002/GenModel" modelDirectory="/tests/src/main/java/"
+    modelPluginID="mapSample"
+    modelName="MapSample" rootExtendsInterface="fr.inria.atlanmod.neoemf.core.PersistentEObject"
+    rootExtendsClass="fr.inria.atlanmod.neoemf.core.impl.PersistentEObjectImpl" reflectiveDelegation="true"
+    importerID="org.eclipse.emf.importer.ecore" bundleManifest="false" featureDelegation="Reflective"
+    complianceLevel="8.0" copyrightFields="false" operationReflection="true" importOrganizing="true"
+    pluginKey="">
+  <foreignModel>MapSample.ecore</foreignModel>
+  <modelPluginVariables>NEOEMF=fr.inria.atlanmod.neoemf.core</modelPluginVariables>
+  <genPackages prefix="MapSample" basePackage="fr.inria.atlanmod.neoemf.tests.models"
+      disposableProviderFactory="true" ecorePackage="MapSample.ecore#/">
+    <genClasses ecoreClass="MapSample.ecore#//SampleModel">
+      <genFeatures createChild="false" ecoreFeature="ecore:EAttribute MapSample.ecore#//SampleModel/name"/>
+      <genFeatures property="None" children="true" createChild="true" ecoreFeature="ecore:EReference MapSample.ecore#//SampleModel/map"/>
+      <genFeatures property="None" children="true" createChild="true" ecoreFeature="ecore:EReference MapSample.ecore#//SampleModel/kvMap"/>
+      <genFeatures property="None" children="true" createChild="true" ecoreFeature="ecore:EReference MapSample.ecore#//SampleModel/contentObjects"/>
+    </genClasses>
+    <genClasses ecoreClass="MapSample.ecore#//StringToStringMap">
+      <genFeatures createChild="false" ecoreFeature="ecore:EAttribute MapSample.ecore#//StringToStringMap/key"/>
+      <genFeatures createChild="false" ecoreFeature="ecore:EAttribute MapSample.ecore#//StringToStringMap/value"/>
+    </genClasses>
+    <genClasses ecoreClass="MapSample.ecore#//K">
+      <genFeatures createChild="false" ecoreFeature="ecore:EAttribute MapSample.ecore#//K/kName"/>
+      <genFeatures createChild="false" ecoreFeature="ecore:EAttribute MapSample.ecore#//K/kInt"/>
+    </genClasses>
+    <genClasses ecoreClass="MapSample.ecore#//V">
+      <genFeatures createChild="false" ecoreFeature="ecore:EAttribute MapSample.ecore#//V/vName"/>
+      <genFeatures createChild="false" ecoreFeature="ecore:EAttribute MapSample.ecore#//V/vInt"/>
+    </genClasses>
+    <genClasses ecoreClass="MapSample.ecore#//KToVMap">
+      <genFeatures notify="false" createChild="false" propertySortChoices="true" ecoreFeature="ecore:EReference MapSample.ecore#//KToVMap/key"/>
+      <genFeatures notify="false" createChild="false" propertySortChoices="true" ecoreFeature="ecore:EReference MapSample.ecore#//KToVMap/value"/>
+    </genClasses>
+    <genClasses ecoreClass="MapSample.ecore#//SampleModelContentObject">
+      <genFeatures createChild="false" ecoreFeature="ecore:EAttribute MapSample.ecore#//SampleModelContentObject/name"/>
+      <genFeatures property="None" notify="false" createChild="false" ecoreFeature="ecore:EReference MapSample.ecore#//SampleModelContentObject/parent"/>
+    </genClasses>
+    <genClasses ecoreClass="MapSample.ecore#//Pack">
+      <genFeatures notify="false" createChild="false" propertySortChoices="true" ecoreFeature="ecore:EReference MapSample.ecore#//Pack/packs"/>
+      <genFeatures notify="false" createChild="false" propertySortChoices="true" ecoreFeature="ecore:EReference MapSample.ecore#//Pack/parentPack"/>
+      <genFeatures property="None" children="true" createChild="true" ecoreFeature="ecore:EReference MapSample.ecore#//Pack/ownedContents"/>
+      <genFeatures createChild="false" ecoreFeature="ecore:EAttribute MapSample.ecore#//Pack/name"/>
+      <genFeatures notify="false" createChild="false" propertySortChoices="true" ecoreFeature="ecore:EReference MapSample.ecore#//Pack/nonContainmentRefComments"/>
+    </genClasses>
+    <genClasses image="false" ecoreClass="MapSample.ecore#//AbstractPackContent">
+      <genFeatures property="None" notify="false" createChild="false" ecoreFeature="ecore:EReference MapSample.ecore#//AbstractPackContent/parentPack"/>
+      <genFeatures createChild="false" ecoreFeature="ecore:EAttribute MapSample.ecore#//AbstractPackContent/name"/>
+      <genFeatures property="None" children="true" createChild="true" ecoreFeature="ecore:EReference MapSample.ecore#//AbstractPackContent/containmentNoOppositeRefComment"/>
+    </genClasses>
+    <genClasses ecoreClass="MapSample.ecore#//PackContent"/>
+    <genClasses ecoreClass="MapSample.ecore#//SpecializedPackContent"/>
+    <genClasses ecoreClass="MapSample.ecore#//PackContent2"/>
+    <genClasses ecoreClass="MapSample.ecore#//AbstractPackContentComment">
+      <genFeatures createChild="false" ecoreFeature="ecore:EAttribute MapSample.ecore#//AbstractPackContentComment/content"/>
+    </genClasses>
+  </genPackages>
+</genmodel:GenModel>
