+++ tests/src/main/resources/MapSample.xcore
@@ -1,7 +1,7 @@
 @Ecore(nsURI="com.mapSample")
 @GenModel(modelDirectory="/test-commons/src/main/java/", complianceLevel="8.0", rootExtendsInterface="fr.inria.atlanmod.neoemf.core.PersistentEObject", rootExtendsClass="fr.inria.atlanmod.neoemf.core.impl.PersistentEObjectImpl", reflectiveDelegation="true", importerID="org.eclipse.emf.importer.ecore", bundleManifest="false", featureDelegation="Reflective")
 
-package fr.inria.atlanmod.neoemf.test.commons.models.mapSample
+package fr.inria.atlanmod.neoemf.tests.models.mapSample
 
 class SampleModel {
 	String name
