+++ tests/src/test/java/fr/inria/atlanmod/neoemf/tests/AllContainmentTest.java
@@ -11,11 +11,11 @@
 
 package fr.inria.atlanmod.neoemf.tests;
 
-import fr.inria.atlanmod.neoemf.test.commons.models.mapSample.AbstractPackContentComment;
-import fr.inria.atlanmod.neoemf.test.commons.models.mapSample.MapSampleFactory;
-import fr.inria.atlanmod.neoemf.test.commons.models.mapSample.MapSamplePackage;
-import fr.inria.atlanmod.neoemf.test.commons.models.mapSample.Pack;
-import fr.inria.atlanmod.neoemf.test.commons.models.mapSample.PackContent;
+import fr.inria.atlanmod.neoemf.tests.models.mapSample.AbstractPackContentComment;
+import fr.inria.atlanmod.neoemf.tests.models.mapSample.MapSampleFactory;
+import fr.inria.atlanmod.neoemf.tests.models.mapSample.MapSamplePackage;
+import fr.inria.atlanmod.neoemf.tests.models.mapSample.Pack;
+import fr.inria.atlanmod.neoemf.tests.models.mapSample.PackContent;
 
 import org.junit.After;
 import org.junit.Before;
